refactor: Separate Page Header Test to a new test file and test the rendering of 'Test Runs' in the Page Header
